Follow the instructions, I created a utils folder inside the src and added web3_funct.js that encaplsulate the loadWeb3, signMessage, loadContract and connect functions respectively. Import these four functions into the App.js and call the functions within the useEffect hook to perform it actions.
